Add wreckage of SpaceShip to level 6.2
Since there's no other SpaceShip wreckage on other Volcano's levels, I think we could add it in this new one to make it looking better, curious and more reasonable addition to game's lore. There's at least one SpaceShip wreckage on every other planet in the game but Volcano. That's very good opportunity to change that.

Also I think that location of the wreckage doesn't collide with anything that is essential for the playability of this level. And placing it in the lava makes that even more curious for future players.
https://cdn.discordapp.com/attachments/808042298700202004/808044752288612402/Schowek01.png